add tracer_db, indexer_db to lib.rs
