+++ evm_loader/cli/src/rpc/db_trx_client.rs
@@ -9,8 +9,8 @@ use solana_sdk::{
     hash::Hash, signature::Signature, transaction::Transaction,
 };
 use solana_transaction_status::{EncodedConfirmedBlock, EncodedConfirmedTransactionWithStatusMeta, TransactionStatus};
-use super::{DbConfig, Rpc, e, tracer_db::TracerDb, indexer_db::IndexerDb};
-use crate::commands::TxParams;
+use super::{Rpc, e,};
+use crate::{types::{tracer_db::TracerDb, indexer_db::IndexerDb, DbConfig, TxParams}};
 use std::any::Any;
 
 #[derive(Debug)]
